Reorganize some files
Move TestComponent out of components folder
Move ThemeProvider into the themes folder
