+++ src/themes/ThemeProvider/index.js
@@ -1,0 +1,7 @@
+// @flow
+
+/* eslint flowtype-errors/enforce-min-coverage: 0 */
+
+import ThemeProvider from './ThemeProvider';
+
+export default ThemeProvider;
